+++ src/knacr/library/validate.py
@@ -2,7 +2,7 @@ from typing import Final, TypeVar
 import re
 
 from jsonschema import ValidationError, validate
-from knacr.container.acr_db import AcrDb
+from knacr.container.acr_db import AcrCoreReg, AcrDb
 from knacr.container.fun.acr_db import check_uri_template, create_acr_db
 from knacr.errors.custom_exceptions import ValJsonEx
 
@@ -24,10 +24,24 @@ def _check_unique(
     return unique_id
 
 
-def _check_changed_to_id(all_ids: set[int], acr_db: AcrDb, /) -> None:
-    for acr_cha in acr_db.acr_changed_to:
-        if acr_cha.id not in all_ids:
+def _check_changed_to_id(cur_acr_con: AcrDb, acr_db: dict[int, AcrDb], /) -> None:
+    if cur_acr_con.deprecated and len(cur_acr_con.acr_synonym) > 0:
+        raise ValJsonEx(
+            f"{cur_acr_con.acr}: 'deprecated' can not have a 'synonyms' field"
+        )
+    if cur_acr_con.deprecated and len(cur_acr_con.acr_changed_to) > 0:
+        raise ValJsonEx(
+            f"{cur_acr_con.acr}: 'deprecated' can not have a 'changed to' field"
+        )
+    for acr_cha in cur_acr_con.acr_changed_to:
+        next_acr_con = acr_db.get(acr_cha.id, None)
+        if next_acr_con is None:
             raise ValJsonEx(f"missing 'changed to' acr id {acr_cha.id}")
+        if next_acr_con.deprecated:
+            raise ValJsonEx(
+                f"{cur_acr_con.acr}: acr can not change into "
+                + f"a deprecated acr {acr_cha.id}"
+            )
 
 
 def _check_missing_link_id(all_ids: set[int], /) -> None:
@@ -45,6 +59,50 @@ def _check_acr_in_reg(acr: str, ccno_reg: str, /) -> None:
     for acr_part in _ACR_SPL.split(acr):
         if acr_part not in ccno_reg:
             raise ValJsonEx(f"{acr} mismatches the acronym in regex: {ccno_reg}")
+
+
+def _check_regex_start_end(reg_full: list[str], reg_part: list[str], /) -> None:
+    for reg in reg_full:
+        if reg[0] != r"^" or reg[-1] != r"$":
+            raise ValJsonEx(f"invalid full regex {reg}")
+    for reg in reg_part:
+        if reg[0] == r"^" or reg[-1] == r"$":
+            raise ValJsonEx(f"invalid part regex {reg}")
+
+
+def _check_list_uniqueness(typ: str, con: list[str], /) -> None:
+    if len(set(con)) != len(con):
+        raise ValJsonEx(f"duplicates in {typ} - {con} found")
+
+
+def _check_pre_suf(typ: str, pr_su: str, pr_su_con: list[str], /) -> None:
+    if pr_su != "" and len(pr_su_con) == 0:
+        raise ValJsonEx(f"{typ} defines {pr_su} but given list is empty")
+    for ps_el in pr_su_con:
+        if ps_el not in pr_su:
+            raise ValJsonEx(f"given list for {typ} has unknown element {ps_el}")
+
+
+def _check_regex(r_ccno: str, r_id: AcrCoreReg, /) -> None:
+    _check_list_uniqueness("prefix", r_id.pre)
+    _check_list_uniqueness("suffix", r_id.suf)
+    _check_regex_start_end([r_ccno, r_id.full], [r_id.core, *r_id.pre, *r_id.suf])
+    if len(r_id.full) <= 2:
+        raise ValJsonEx(f"regex for id must be longer than 2 {r_id.full}")
+    if r_id.full[1:] not in r_ccno:
+        raise ValJsonEx(
+            f"regex for ccno must contain regex for id: {r_id.full} -> {r_ccno}"
+        )
+    if (
+        pre_suf := re.compile(rf"^(.*){re.escape(r_id.core)}(.*)$").match(r_id.full)
+    ) is not None:
+        pre, suf = pre_suf.groups()
+        _check_pre_suf("prefix", pre[1:], r_id.pre)
+        _check_pre_suf("suffix", suf[0:-2], r_id.suf)
+    else:
+        raise ValJsonEx(
+            f"regex for id must contain regex for core: {r_id.core} -> {r_id.full}"
+        )
 
 
 def _check_loops(cur: AcrDb, full: dict[int, AcrDb], ids: set[int], /) -> None:
@@ -65,13 +123,14 @@ def _validate_acr_db_dc(to_eval_acr: dict[int, AcrDb], /) -> None:
     for acr_id, acr_db in to_eval_acr.items():
         check_uri_template(acr_db.catalogue)
         uniqueness.add(_check_unique(uniqueness, acr_db))
-        _check_changed_to_id(all_ids, acr_db)
+        _check_changed_to_id(acr_db, to_eval_acr)
         _check_acr(acr_db.acr)
+        _check_regex(acr_db.regex_ccno, acr_db.regex_id)
         _check_acr_in_reg(acr_db.acr, acr_db.regex_ccno)
         _check_loops(acr_db, to_eval_acr, {acr_id})
 
 
-def validate_acr_db_schema(to_eval: _TJ, /) -> None:
+def validate_acr_db(to_eval: _TJ, /) -> dict[int, AcrDb]:
     if not isinstance(to_eval, dict):
         raise ValJsonEx(f"expected a dictionary, got {type(to_eval)}")
     try:
@@ -83,6 +142,8 @@ def validate_acr_db_schema(to_eval: _TJ, /) -> None:
         raise ValJsonEx(
             f"Acronym Data is incorrectly formatted! [{exc.message}]"
         ) from exc
+    else:
+        return acr_db
 
 
 def validate_min_acr_db_schema(to_eval: _TJ, /) -> None: